Status Effects! (And other small things)

+++ Fakemon/src/fakemon/BattleAction.java
@@ -1,7 +1,7 @@
 package fakemon;
 
 public abstract class BattleAction {
-	public abstract int getSpeed();
+	public abstract float getSpeed();
 	public abstract int getPriority();
 	public abstract void doAction(BattleScreen bsc);
 	public abstract boolean validate();
